Removed the generic ajax loader
