--- src/constants/ajaxLoader/genericAjaxLoaderActionTypes.js
@@ -1,6 +1,0 @@
-const genericAjaxLoaderActionTypes = {
-  HIDE: 'HIDE_GENERIC_AJAX_LOADER',
-  SHOW: 'SHOW_GENERIC_AJAX_LOADER'
-};
-
-export default genericAjaxLoaderActionTypes;
